ART: Add new profile saver options
save without jit & profile AOT code

--Xps-save-without-jit-notifications
The hotness for system server code is increased in AOT-ed code. The flow
does not call into JIT and as such notifications are not triggered.
Instead of relying on the JIT system, make use of a simple back off
strategy to save the profile.

--Xps-profile-aot-code
Starts the profile saver even if the oat file is speed compiled.

Test: m test-art-host
      boot a device with system server profiling enabled.
Bug: 73313191

Change-Id: I66ec422a76bc9244349da7a5d18a3df5bcc9ccb7
Signed-off-by: mydongistiny <jaysonedson@gmail.com>