Move InterpreterCache to right below tlsPtr_
Change the position of the InterpreterCache field in Thread to be
directly below the tlsPtr_ field. Since both members of the tlsPtr_
and InterpreterCache fields are used by asm_code we need their offsets
in asm_support.h. The fields at the end of the Thread struct have been
undergoing changes. By moving this field up we avoid the need to
update asm_support.h whenever one of the fields is modified.

Test: ./test.py --host
Change-Id: Ic2863116ed446af155badfc3bf098add7ba0b699
Signed-off-by: mydongistiny <jaysonedson@gmail.com>