Remove bad code from ComputeRelativeTimeSpec
The test in the "else if" clause was off by one. But it should never be
exercised anyway. Replace it with a correct DCHECK.
Have ComputeRelativeTimeSpec return false, rather than true, on failure
to compute a valid timespec.

Update num_contenders if lock acquisition results in unexpected
failure. This matters only in that it may put us into a slightly
better state while logging debug information.

Fix an unrelated comment typo.

Bug: 238032384
Test: Build and boot AOSP eng build
Change-Id: Idf3a0b4c349168370079d24457ea134b06cf231e
Signed-off-by: Jason Edson <jaysonedson@gmail.com>